+++ app/src/test/kotlin/kt/school/starlord/ui/categories/CategoriesViewModelTest.kt
@@ -5,6 +5,7 @@ import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
 import kt.school.starlord.entity.Category
+import kt.school.starlord.entity.Subcategory
 import kt.school.starlord.model.network.NetworkRepository
 import kt.school.starlord.model.room.RoomRepository
 import kt.school.starlord.ui.TestCoroutineRule
@@ -28,7 +29,15 @@ internal class CategoriesViewModelTest {
     @Test
     fun `load categories with subcategories from network`() = testCoroutineRule.runBlockingTest {
         // Given
-        val categoriesWithSubcategories = MockedData.categoriesWithSubcategories
+        val categoriesWithSubcategories = mapOf(
+            Category("categoryName3") to listOf(
+                Subcategory("subcategoryName1", "categoryName3", 5, "link1"),
+                Subcategory("subcategoryName2", "categoryName3", 2, "link2")
+            ),
+            Category("categoryName4") to listOf(
+                Subcategory("subcategoryName3", "categoryName4", 3, "link3")
+            )
+        )
         coEvery { networkRepository.getCategoriesWithSubcategories() }.coAnswers { categoriesWithSubcategories }
 
         // When
@@ -48,7 +57,7 @@ internal class CategoriesViewModelTest {
     @Test
     fun `load categories from database`() = testCoroutineRule.runBlockingTest {
         // Given
-        val categories: List<Category> = MockedData.categories
+        val categories = listOf(Category("categoryName1"), Category("categoryName2"))
         coEvery { roomRepository.getCategories() }.coAnswers { categories }
 
         // When
